fix(quality_score) (ludo,nicholas) quality score should be a decimal instead of a integer

+++ app/modules/main/services/creatives.js
@@ -57,7 +57,7 @@ angular
                 name: "Quality Score",
                 shortName: "QSCORE",
                 description: "How much google loves you",
-                format: "integer"
+                format: "decimal"
             }
         };
     });
